Allow jobs to be pushed synchronously
This introduces the ability for jobs to be pushed synchronously using the option
`skip_pipeline: true`. This is necessary when pushing jobs as part of a batch to
ensure they are submitted prior to the batch being committed--without this,
there's the possibility that jobs for the batch are left on the PushPipeline
when the batch is committed, and those jobs are not included in the batch.

As part of this work, the logic for pushing jobs was refactored so that it lives
in the `FaktoryWorker.Job` module and `FaktoryWorker.PushPipeline.Consumer`
relies on the `FaktoryWorker.Job.push/2` function to push work.