Fix inconsistent grabbing of parent_bid (was mistakenly parent_id)
This is the result of some debugging today that I had to do.

The situation was basically this:

Job A fires off Job B

Job B creates a batch, and passes its big to another module that will
run 4 jobs in their own batches.

The original batch is re-opened to push the new jobs.

We pass the configuration to these new batches to let them know who
their parent is.

In the `Map.put` we put the keyword to `parent_bid`, however in the opts
for the function, we look for `parent_id`.

The reason this came to light is my success callbacks were firing while
longer-running jobs down the chain were still not truly complete.

With this, the API should be consistent. I imagine it was a typo
originally.