Add warn/error level logging in telemetry to appropriate events
This comes in response to some problems we had in production. Ideally,
we should have warning/error level logs for certain events.

This does the following:

- log warn when `NOTUNIQUE`
- log warn on heartbeat failure
- log error on job failure
- log warn on failed `ACK` with `:ok` status
- log error on failed `ACK` with `:error` status

It maintains the same formatting as the `info` level helper already set
up, as to not mess up any saved searches folks may have, but gives this
granularity for easier inspection in situations like mass job-failure.