add :retry_on_error config option
This option will allow users to opt-in to a new behaviour, where the job's return value will be considered when determing whether to retry the job. Previously, only a `raise` would constitute a retry-able failure. With this new behaviour enabled, returning `:error` or `{:error, term()}` will also trigger a retry.